+++ Restaurant/Restaurant/Classes/ApiClasses/BranchStore.cs
@@ -18,14 +18,14 @@ namespace Restaurant.Classes
 
     class BranchStoretable
     {
-        public int id { get; set; }
-        public Nullable<int> branchId { get; set; }
-        public Nullable<int> storeId { get; set; }
+        public long id { get; set; }
+        public Nullable<long> branchId { get; set; }
+        public Nullable<long> storeId { get; set; }
         public string notes { get; set; }
         public Nullable<System.DateTime> createDate { get; set; }
         public Nullable<System.DateTime> updateDate { get; set; }
-        public Nullable<int> createUserId { get; set; }
-        public Nullable<int> updateUserId { get; set; }
+        public Nullable<long> createUserId { get; set; }
+        public Nullable<long> updateUserId { get; set; }
         public int isActive { get; set; }
         public Boolean canDelete { get; set; } 
 
@@ -33,21 +33,21 @@ namespace Restaurant.Classes
     }
     class BranchStore
     {
-        public int id { get; set; }
-        public Nullable<int> branchId { get; set; }
-        public Nullable<int> storeId { get; set; }
+        public long id { get; set; }
+        public Nullable<long> branchId { get; set; }
+        public Nullable<long> storeId { get; set; }
         public string notes { get; set; }
         public Nullable<System.DateTime> createDate { get; set; }
         public Nullable<System.DateTime> updateDate { get; set; }
-        public Nullable<int> createUserId { get; set; }
-        public Nullable<int> updateUserId { get; set; }
+        public Nullable<long> createUserId { get; set; }
+        public Nullable<long> updateUserId { get; set; }
         public int isActive { get; set; }
 
 
         public Boolean canDelete { get; set; }
 
         // branch
-        public int bbranchId { get; set; }
+        public long bbranchId { get; set; }
         public string bcode { get; set; }
         public string bname { get; set; }
         public string baddress { get; set; }
@@ -56,15 +56,15 @@ namespace Restaurant.Classes
         public string bmobile { get; set; }
         public Nullable<System.DateTime> bcreateDate { get; set; }
         public Nullable<System.DateTime> bupdateDate { get; set; }
-        public Nullable<int> bcreateUserId { get; set; }
-        public Nullable<int> bupdateUserId { get; set; }
+        public Nullable<long> bcreateUserId { get; set; }
+        public Nullable<long> bupdateUserId { get; set; }
         public string bnotes { get; set; }
-        public Nullable<int> bparentId { get; set; }
+        public Nullable<long> bparentId { get; set; }
         public byte bisActive { get; set; }
         public string btype { get; set; }
 
         // store
-        public int sbranchId { get; set; }
+        public long sbranchId { get; set; }
         public string scode { get; set; }
         public string sname { get; set; }
         public string saddress { get; set; }
@@ -73,10 +73,10 @@ namespace Restaurant.Classes
         public string smobile { get; set; }
         public Nullable<System.DateTime> screateDate { get; set; }
         public Nullable<System.DateTime> supdateDate { get; set; }
-        public Nullable<int> screateUserId { get; set; }
-        public Nullable<int> supdateUserId { get; set; }
+        public Nullable<long> screateUserId { get; set; }
+        public Nullable<long> supdateUserId { get; set; }
         public string snotes { get; set; }
-        public Nullable<int> sparentId { get; set; }
+        public Nullable<long> sparentId { get; set; }
         public byte sisActive { get; set; }
         public string stype { get; set; }
          
@@ -99,7 +99,7 @@ namespace Restaurant.Classes
        
            
         // get is exist
-        public async Task<int> UpdateStoresById(List<BranchStoretable> newList, int branchId, int userId)
+        public async Task<int> UpdateStoresById(List<BranchStoretable> newList, long branchId, long userId)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "BranchStore/UpdateStoresById";
