+++ Restaurant/Restaurant/Classes/ApiClasses/BranchesUsers.cs
@@ -17,31 +17,31 @@ namespace Restaurant.Classes
 
     public class BranchesUserstable
     {
-        public int branchsUsersId { get; set; }
-        public Nullable<int> branchId { get; set; }
-        public Nullable<int> userId { get; set; }
+        public long branchsUsersId { get; set; }
+        public Nullable<long> branchId { get; set; }
+        public Nullable<long> userId { get; set; }
         public Nullable<System.DateTime> createDate { get; set; }
         public Nullable<System.DateTime> updateDate { get; set; }
-        public Nullable<int> createUserId { get; set; }
-        public Nullable<int> updateUserId { get; set; }
+        public Nullable<long> createUserId { get; set; }
+        public Nullable<long> updateUserId { get; set; }
          
     }
 
     public class BranchesUsers
     {
-        public int branchsUsersId { get; set; }
-        public Nullable<int> branchId { get; set; }
-        public Nullable<int> userId { get; set; }
+        public long branchsUsersId { get; set; }
+        public Nullable<long> branchId { get; set; }
+        public Nullable<long> userId { get; set; }
         public Nullable<System.DateTime> createDate { get; set; }
         public Nullable<System.DateTime> updateDate { get; set; }
-        public Nullable<int> createUserId { get; set; }
-        public Nullable<int> updateUserId { get; set; }
+        public Nullable<long> createUserId { get; set; }
+        public Nullable<long> updateUserId { get; set; }
 
 
 
 
         // branch
-        public int bbranchId { get; set; }
+        public long bbranchId { get; set; }
         public string bcode { get; set; }
         public string bname { get; set; }
         public string baddress { get; set; }
@@ -50,15 +50,15 @@ namespace Restaurant.Classes
         public string bmobile { get; set; }
         public Nullable<System.DateTime> bcreateDate { get; set; }
         public Nullable<System.DateTime> bupdateDate { get; set; }
-        public Nullable<int> bcreateUserId { get; set; }
-        public Nullable<int> bupdateUserId { get; set; }
+        public Nullable<long> bcreateUserId { get; set; }
+        public Nullable<long> bupdateUserId { get; set; }
         public string bnotes { get; set; }
-        public Nullable<int> bparentId { get; set; }
+        public Nullable<long> bparentId { get; set; }
         public byte bisActive { get; set; }
         public string btype { get; set; }
 
         // user
-        public int uuserId { get; set; }
+        public long uuserId { get; set; }
         public string uusername { get; set; }
         public string upassword { get; set; }
         public string uname { get; set; }
@@ -67,8 +67,8 @@ namespace Restaurant.Classes
         public string uworkHours { get; set; }
         public DateTime? ucreateDate { get; set; }
         public DateTime? uupdateDate { get; set; }
-        public int? ucreateUserId { get; set; }
-        public int? uupdateUserId { get; set; }
+        public long? ucreateUserId { get; set; }
+        public long? uupdateUserId { get; set; }
         public string uphone { get; set; }
         public string umobile { get; set; }
         public string uemail { get; set; }
@@ -80,7 +80,7 @@ namespace Restaurant.Classes
         public string uimage { get; set; }
       
        
-        public async Task<List<BranchesUsers>> GetBranchesByUserId(int userId)
+        public async Task<List<BranchesUsers>> GetBranchesByUserId(long userId)
         {
             List<BranchesUsers> items = new List<BranchesUsers>();
             Dictionary<string, string> parameters = new Dictionary<string, string>();
@@ -99,7 +99,7 @@ namespace Restaurant.Classes
             return items;
         }
         
-        public async Task<int> UpdateBranchByUserId(List<BranchesUserstable> newList, int userId, int updateUserId)
+        public async Task<int> UpdateBranchByUserId(List<BranchesUserstable> newList, long userId, long updateUserId)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "BranchesUsers/UpdateBranchByUserId";
